requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==2.3.7
paho-mqtt==1.6.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
